# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Talk API
  description: |
    Text-to-speech API designed for AI agents. 
    Send text, receive audio — one POST request, instant voice.
    
    ## Authentication
    
    - **API Key Auth**: For TTS operations (`Authorization: Bearer YOUR_API_KEY`)
    - **Session Auth**: For user account management (cookie-based)
    
    ## Quick Start
    
    ```bash
    # Sign up and get an API key
    curl -X POST https://talk.onhyper.io/api/v1/auth/signup \
      -H "Content-Type: application/json" \
      -d '{"email": "you@example.com", "password": "your-password"}'
    
    # Convert text to speech
    curl -X POST https://talk.onhyper.io/api/v1/memo \
      -H "Authorization: Bearer YOUR_API_KEY" \
      -H "Content-Type: application/json" \
      -d '{"text": "Hello!", "voice": "rachel"}'
    ```
  version: 1.0.0
  contact:
    name: Agent Talk Support
    email: hello@hyper.io
    url: https://talk.onhyper.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://talk.onhyper.io
    description: Production
  - url: http://localhost:3001
    description: Development
...
paths:
...
  /api/keys:
    get:
      tags: [API Keys]
      summary: List API keys
      description: List all API keys for the authenticated user
      operationId: listApiKeys
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [API Keys]
      summary: Create an API key
      description: Create a new API key
      operationId: createApiKey
      security:
        - cookieAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateApiKeyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/keys/{keyId}:
    parameters:
      - name: keyId
        in: path
        required: true
        schema:
          type: string
        description: API key ID

    get:
      tags: [API Keys]
      summary: Get an API key
      description: Get a specific API key by ID
      operationId: getApiKey
      security:
        - cookieAuth: []
      responses:
        '200':
          description: API key details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [API Keys]
      summary: Revoke an API key
      description: Revoke (soft delete) an API key
      operationId: revokeApiKey
      security:
        - cookieAuth: []
      responses:
        '200':
          description: API key revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [API Keys]
      summary: Update an API key
      description: Update an API key (currently only name)
      operationId: updateApiKey
      security:
        - cookieAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateApiKeyRequest'
      responses:
        '200':
          description: API key updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================
  # Text-to-Speech
  # ============================================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key authentication
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: Session cookie authentication

  # ============================================
  # Responses
  # ============================================
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
        prefix:
          type: string
        maskedKey:
          type: string
        name:
          type: string
          nullable: true
        usageCount:
          type: integer
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ApiKeyFull:
      type: object
      properties:
        id:
          type: string
        key:
          type: string
        prefix:
          type: string
        name:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    ApiKeyListResponse:
      type: object
      properties:
        success:
          type: boolean
        keys:
          type: array
          items:
            $ref: '#/components/schemas/ApiKey'
    ApiKeyResponse:
      type: object
      properties:
        success:
          type: boolean
        key:
          $ref: '#/components/schemas/ApiKey'
    CreateApiKeyRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        test:
          type: boolean
          default: false
    CreateApiKeyResponse:
      type: object
      properties:
        success:
          type: boolean
        key:
          $ref: '#/components/schemas/ApiKeyFull'
        warning:
          type: string
    UpdateApiKeyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 100

    # ---- Voice ----
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - INVALID_INPUT
                - MISSING_FIELD
                - INVALID_VOICE
                - UNAUTHORIZED
                - MISSING_API_KEY
                - INVALID_API_KEY
                - EXPIRED_TOKEN
                - FORBIDDEN
                - INSUFFICIENT_TIER
                - REVOKED_KEY
                - NOT_FOUND
                - MEMO_NOT_FOUND
                - RATE_LIMIT_EXCEEDED
                - DAILY_LIMIT_EXCEEDED
                - MONTHLY_LIMIT_EXCEEDED
                - INTERNAL_ERROR
                - TTS_SERVICE_ERROR
                - STORAGE_ERROR
                - NOT_IMPLEMENTED
                - SERVICE_UNAVAILABLE
            message:
              type: string
            details:
              type: object
              additionalProperties: true